Moved MuxedAccount type to a separate file.
